refactor: name the module graph_codec (was structural)
The module is the codec between a Vizij GraphSpec and the Arora shared Graph;
"structural" only distinguished it from the opaque spec-carrier it replaced.
"structural" survives as the adjective for the encoding it produces.
