+++ crates/interop/vizij-arora-behavior/src/graph_codec.rs
